Feature/#296 dave 2025 migration (#299)

+++ src/F1.Infrastructure/Data/Entities/MigrationImportParticipantDeltaSummaryEntity.cs
@@ -6,8 +6,8 @@ public sealed class MigrationImportParticipantDeltaSummaryEntity
     public Guid ImportRunId { get; set; }
     public string Subject { get; set; } = string.Empty;
     public int ImportedTotalPoints { get; set; }
-    public int CalculatedTotalPoints { get; set; }
-    public int NetDeltaPoints { get; set; }
+    public decimal CalculatedTotalPoints { get; set; }
+    public decimal NetDeltaPoints { get; set; }
     public string? TopReasonCode { get; set; }
     public int TopReasonCount { get; set; }
 }
